+++ rocksdb/interfaces.py
@@ -42,21 +42,6 @@ class MergeOperator:
         pass
 
 
-class FilterPolicy:
-    __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def name(self):
-        pass
-
-    @abstractmethod
-    def create_filter(self, keys):
-        pass
-
-    @abstractmethod
-    def key_may_match(self, key, filter_):
-        pass
-
 class SliceTransform:
     __metaclass__ = ABCMeta